genome-guide/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
biopython==1.84